Base pattern applied.
