+++ SignumGenerator/SignumLib/Figures/IPattern.cs
@@ -4,7 +4,12 @@ using System.Drawing;
 
 namespace SignumLib.Figures
 {
-    public interface IPattern
+    public interface IPatternBase
+    {
+        void Draw(Graphics g, SignumData data, InputBaseData input);
+    }
+
+    public interface IPatternLayer
     {
         void Draw(Graphics g, SignumData data, InputLayerData input);
     }
